Remove style group nodes and refactor node type enum
