+++ src/node/Style.h
@@ -6,127 +6,7 @@
 
 namespace chtl {
 
-/**
- * 样式组节点
- * 表示一组相关的样式定义
- */
-class StyleGroupNode : public Node {
-public:
-    explicit StyleGroupNode(const std::string& name = "",
-                           const NodePosition& position = NodePosition());
-    virtual ~StyleGroupNode() = default;
-    
-    // 样式组名称
-    const std::string& getGroupName() const;
-    void setGroupName(const std::string& name);
-    
-    // 是否为值样式组（有具体值）
-    bool isValuedGroup() const;
-    void setValuedGroup(bool valued);
-    
-    // 样式管理
-    void addStyleProperty(const std::string& property, const std::string& value);
-    void removeStyleProperty(const std::string& property);
-    std::string getStyleProperty(const std::string& property) const;
-    std::unordered_map<std::string, std::string> getAllStyleProperties() const;
-    
-    // CSS规则管理
-    void addCssRule(std::shared_ptr<CssRuleNode> rule);
-    void removeCssRule(const std::string& selector);
-    std::vector<std::shared_ptr<CssRuleNode>> getCssRules() const;
-    std::shared_ptr<CssRuleNode> findCssRule(const std::string& selector) const;
-    
-    // 嵌套样式组管理
-    void addNestedGroup(std::shared_ptr<StyleGroupNode> group);
-    void removeNestedGroup(const std::string& groupName);
-    std::vector<std::shared_ptr<StyleGroupNode>> getNestedGroups() const;
-    std::shared_ptr<StyleGroupNode> findNestedGroup(const std::string& groupName) const;
-    
-    // 继承管理
-    void addInheritedGroup(const std::string& groupName);
-    void removeInheritedGroup(const std::string& groupName);
-    std::vector<std::string> getInheritedGroups() const;
-    
-    // 验证
-    bool validate() const override;
-    
-    // 序列化
-    std::string toString() const override;
-    std::string toDebugString(int indent = 0) const override;
-    std::string toCss() const;
-    
-    // 克隆
-    std::shared_ptr<Node> clone() const override;
-    
-    // 访问者模式
-    void accept(NodeVisitor& visitor) override;
-    
-private:
-    std::string groupName_;
-    bool isValuedGroup_;
-    std::unordered_map<std::string, std::string> styleProperties_;
-    std::vector<std::shared_ptr<CssRuleNode>> cssRules_;
-    std::vector<std::shared_ptr<StyleGroupNode>> nestedGroups_;
-    std::vector<std::string> inheritedGroups_;
-};
 
-/**
- * 样式组调用节点
- * 表示对样式组的调用和使用
- */
-class StyleGroupCallNode : public Node {
-public:
-    explicit StyleGroupCallNode(const std::string& groupName = "",
-                               const NodePosition& position = NodePosition());
-    virtual ~StyleGroupCallNode() = default;
-    
-    // 样式组名称
-    const std::string& getGroupName() const;
-    void setGroupName(const std::string& groupName);
-    
-    // 调用类型
-    enum class CallType {
-        DIRECT,        // 直接调用
-        INHERITED,     // 继承调用
-        SPECIALIZED    // 特例化调用
-    };
-    
-    CallType getCallType() const;
-    void setCallType(CallType type);
-    
-    // 参数管理（用于特例化）
-    void addParameter(const std::string& name, const std::string& value);
-    void removeParameter(const std::string& name);
-    std::unordered_map<std::string, std::string> getParameters() const;
-    bool hasParameter(const std::string& name) const;
-    std::string getParameterValue(const std::string& name) const;
-    
-    // 覆盖属性管理
-    void addOverride(const std::string& property, const std::string& value);
-    void removeOverride(const std::string& property);
-    std::unordered_map<std::string, std::string> getOverrides() const;
-    
-    // 验证
-    bool validate() const override;
-    
-    // 序列化
-    std::string toString() const override;
-    std::string toDebugString(int indent = 0) const override;
-    
-    // 克隆
-    std::shared_ptr<Node> clone() const override;
-    
-    // 访问者模式
-    void accept(NodeVisitor& visitor) override;
-    
-private:
-    std::string groupName_;
-    CallType callType_;
-    std::unordered_map<std::string, std::string> parameters_;
-    std::unordered_map<std::string, std::string> overrides_;
-    
-    std::string callTypeToString() const;
-};
 
 /**
  * 伪类节点
